added function to fetch all registrations from firestore on startup

+++ src/internal/firebase_client/bundle_update.go
@@ -3,7 +3,7 @@ package firebase_client
 func NewBundledUpdate() *BundledUpdate {
 	return &BundledUpdate{
 		Ready:           false,
-		InvocationCount: make(map[string]int),
+		InvocationCount: make(map[string]int64),
 		Registrations:   make(map[string]RegistrationAction),
 		Cache:           make(map[string]string),
 	}
